Replace app 'kind' strings with AppDefinition.
This separates the concept of an App and a RemoteApp.  There is only one
Client Hub, but it's installed a hundred times depending on how many
clients exist.  This makes sense, but it was confusing initially.

I think this helps restrict what RemoteApp has to know about the
individual AppDefinitions, and removes most of the  hardcoded references
in the code.  Places like Entry#find_or_create_from_hentry previously
had to know the individual kinds that each client needed.  Rather than
just knowing the kinds, we can set some metadata on the AppDefinition
and programatically determine which should be spun up.

I started down this path because my next piece of work is to add a
namespace environment variable that will replace the hardcoded "g5-" in
front of everything.  I didn't like the way I was going to have to
implement it under the previous setup.